+++ src/TtrayIcon.cpp
@@ -271,7 +271,7 @@ HMENU TtrayIconDec::createMenu(int &ord)
      for (comptr<IBaseFilter> bff;eff->Next(1,&bff,NULL)==S_OK;bff=NULL)
       {
        char_t name[MAX_PATH],filtername[MAX_PATH];
-       getFilterName(bff,name,filtername,(int)countof(filtername));
+       getFilterName(bff,name,filtername,countof(filtername));
        char_t compactname[41];
        PathCompactPathEx(compactname,name,40,0);
        graphnames[insertMenuItem(hm,ord,IDC_FIRST_GRAPH+(i++),compactname,false,false,!!comptrQ<ISpecifyPropertyPages>(bff))]=filtername;
